# backend/src/swagger.yaml
openapi: 3.0.3
info:
  title: UAIFood API
  version: 1.0.0
  description: |
    API do UAIFood (YAML). Segurança JWT via Bearer JWT.
    Também é aceito o cabeçalho `x-access-token` como alternativa (estilo jwt-example).

    ## Autenticação
    Esta API exige autenticação para a maioria dos endpoints. Você pode enviar o token JWT de duas formas:

    1) Header Authorization (preferencial)

    ```bash
    curl -H "Authorization: Bearer <seu_token>" http://localhost:3333/api/users/me
    ```

    2) Header x-access-token (compatibilidade com jwt-example)

    ```bash
    curl -H "x-access-token: <seu_token>" http://localhost:3333/api/users/me
    ```

    Para obter o token, faça login em `/api/auth/login`.
servers:
  - url: http://localhost:3333
    description: Desenvolvimento local
security:
  # Requisitos globais: qualquer endpoint exige autenticação (Bearer ou x-access-token),
  # exceto aqueles que explicitamente definirem `security: []` nas suas operações abaixo.
  - bearerAuth: []
  - xAccessToken: []
...
paths:
...
  /api/orders:
    get:
      tags: [Orders]
      summary: Lista pedidos do usuário autenticado
      security:
        - bearerAuth: []
        - xAccessToken: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  orders:
                    type: array
                    items: { $ref: '#/components/schemas/OrderWithItems' }
              examples:
                exemplo:
                  value:
                    orders:
                      - id: ord-123
                        status: PENDENTE
                        table_number: '5'
                        payment_method: PIX
                        total: 33.9
                        created_at: '2025-11-12T12:00:00Z'
                        items:
                          - id: oi-1
                            quantity: 2
                            unit_price: 6.5
                            total_price: 13.0
                            item:
                              id: it-coca
                              name: Coca-Cola Lata
                              price: 6.5
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Orders]
      summary: Cria pedido (mesa obrigatória)
      security:
        - bearerAuth: []
        - xAccessToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateOrderRequest' }
            examples:
              exemplo:
                value:
                  tableNumber: '7'
                  paymentMethod: PIX
                  items:
                    - itemId: it-coca
                      quantity: 2
                    - itemId: it-burguer
                      quantity: 1
      responses:
        '201': { description: Criado }
        '400': { $ref: '#/components/responses/BadRequest' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /api/admin/orders:
    get:
      tags: [Admin]
      summary: Lista pedidos (ADMIN/ROOT)
      security:
        - bearerAuth: []
        - xAccessToken: []
      parameters:
        - in: query
          name: status
          schema: { $ref: '#/components/schemas/OrderStatus' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  orders:
                    type: array
                    items: { $ref: '#/components/schemas/OrderWithItemsAdmin' }
              examples:
                exemplo:
                  value:
                    orders:
                      - id: ord-123
                        status: PREPARANDO
                        table_number: '5'
                        payment_method: PIX
                        total: 33.9
                        created_at: '2025-11-12T12:00:00Z'
                        user:
                          id: uuid-user
                          name: Maria Silva
                          email: maria@exemplo.com
                        items:
                          - id: oi-1
                            quantity: 2
                            unit_price: 6.5
                            total_price: 13.0
                            item:
                              id: it-coca
                              name: Coca-Cola Lata
                              price: 6.5
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/admin/orders/{id}/status:
    patch:
      tags: [Admin]
      summary: Atualiza status do pedido
      security:
        - bearerAuth: []
        - xAccessToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SetOrderStatusRequest' }
            examples:
              exemplo:
                value:
                  status: PRONTO
      responses:
        '200': { description: OK }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/admin/orders/{id}:
    delete:
      tags: [Admin]
      summary: Cancela pedido
      security:
        - bearerAuth: []
        - xAccessToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '204': { description: Cancelado }
        '404': { $ref: '#/components/responses/NotFound' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    xAccessToken:
      type: apiKey
      in: header
      name: x-access-token
      schemas:
    Error:
      type: object
      required: [ok, message]
      properties:
        ok: { type: boolean, example: false }
        message: { type: string, example: 'Ocorreu um erro.' }
        code: { type: string, nullable: true, example: 'UNAUTHORIZED' }
        details:
          type: object
          nullable: true
          additionalProperties: true
    ValidationError:
      type: object
      required: [ok, message, errors]
      properties:
        ok: { type: boolean, example: false }
        message: { type: string, example: 'Dados inválidos.' }
        errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string, example: 'email' }
              message: { type: string, example: 'Email inválido.' }
        RegisterRequest:
          type: object
          required: [name, email, password]
          properties:
            name: { type: string }
            email: { type: string, format: email }
            password: { type: string, format: password }
...